HW2_v2

+++ src/main/java/ru/javawebinar/topjava/repository/InMemoryUserMealRepository.java
@@ -19,31 +19,46 @@ public class InMemoryUserMealRepository implements UserMealRepository {
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 30, 10, 0), "Завтрак", 500, 1));
-        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 30, 13, 0), "Обед", 1000, 1));
-        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 30, 20, 0), "Ужин", 500, 1));
-        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 10, 0), "Завтрак", 1000, 1));
-        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 13, 0), "Обед", 500, 1));
-        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 20, 0), "Ужин", 501, 1));
-        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 10, 0), "Завтрак (чужой)", 501, 2));
+        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 30, 10, 0), "Завтрак", 500, 1), 1);
+        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 30, 13, 0), "Обед", 1000, 1), 1);
+        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 30, 20, 0), "Ужин", 500, 1), 1);
+        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 10, 0), "Завтрак", 1000, 1), 1);
+        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 13, 0), "Обед", 500, 1), 1);
+        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 20, 0), "Ужин", 501, 1), 1);
+        save(new UserMeal(LocalDateTime.of(2015, Month.MAY, 31, 10, 0), "Завтрак (чужой)", 501, 2), 2);
     }
 
     @Override
-    public UserMeal save(UserMeal userMeal) {
+    public UserMeal save(UserMeal userMeal, int userId) {
         if (userMeal.isNew()) {
             userMeal.setId(counter.incrementAndGet());
+        } else {
+            UserMeal persistedMeal = get(userMeal.getId(), userId);
+            if (persistedMeal == null) {
+                return null;
+            }
         }
+        userMeal.setUserId(userId);
         return repository.put(userMeal.getId(), userMeal);
     }
 
     @Override
-    public void delete(int id) {
+    public boolean delete(int id, int userId) {
+        UserMeal meal = get(id, userId);
+        if (meal == null) {
+            return false;
+        }
         repository.remove(id);
+        return true;
     }
 
     @Override
-    public UserMeal get(int id) {
-        return repository.get(id);
+    public UserMeal get(int id, int userId) {
+        UserMeal result = repository.get(id);
+        if (result.getUserId() != userId) {
+            return null;
+        }
+        return result;
     }
 
     @Override
